feat: rename ocean text list base
