+++ app/src/main/java/br/com/useblu/oceands/client/ui/textlistitem/TextListItemActivity.kt
@@ -1,4 +1,4 @@
-package br.com.useblu.oceands.client.ui.textlistitembase
+package br.com.useblu.oceands.client.ui.textlistitem
 
 import android.os.Bundle
 import android.widget.Toast
@@ -6,19 +6,19 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.ViewModelProvider
 import br.com.useblu.oceands.client.R
-import br.com.useblu.oceands.client.databinding.ActivityTextListItemBaseBinding
+import br.com.useblu.oceands.client.databinding.ActivityTextListItemBinding
 
-class TextListItemBaseActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityTextListItemBaseBinding
-    private lateinit var viewModel: TextListItemBaseViewModel
+class TextListItemActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityTextListItemBinding
+    private lateinit var viewModel: TextListItemViewModel
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_text_list_item_base)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_text_list_item)
         binding.lifecycleOwner = this
 
-        viewModel = ViewModelProvider(this)[TextListItemBaseViewModel::class.java]
+        viewModel = ViewModelProvider(this)[TextListItemViewModel::class.java]
         binding.viewmodel = viewModel
         initObservers()
     }
