+++ app/src/main/java/br/com/useblu/oceands/client/ui/textlistitem/TextListItemViewModel.kt
@@ -1,10 +1,10 @@
-package br.com.useblu.oceands.client.ui.textlistitembase
+package br.com.useblu.oceands.client.ui.textlistitem
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class TextListItemBaseViewModel : ViewModel() {
+class TextListItemViewModel : ViewModel() {
     private val _touchItem: MutableLiveData<String> = MutableLiveData()
     val touchItem: LiveData<String> get() = _touchItem
