+++ ocean-components/src/main/res/layout/ocean_expandable_text_list_icon_item.xml
